+++ src/main/java/com/project/ProjectApplication.java
@@ -13,21 +13,6 @@ import java.util.Locale;
 public class ProjectApplication {
 
 
-    @Bean
-    public LocaleResolver localeResolver() {
-        SessionLocaleResolver slr = new SessionLocaleResolver();
-        slr.setDefaultLocale(Locale.US);
-        return slr;
-    }
-
-
-    @Bean
-    public LocaleChangeInterceptor localeChangeInterceptor() {
-        LocaleChangeInterceptor lci = new LocaleChangeInterceptor();
-        lci.setParamName("lang");
-        return lci;
-    }
-
 
 
     public static void main(String[] args) {
