+++ src/main/java/com/project/springConfig/webConfig.java
@@ -1,22 +1,36 @@
 package com.project.springConfig;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.Locale;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-
+import org.springframework.web.servlet.i18n.SessionLocaleResolver;
 
 @Configuration
+@ComponentScan(basePackages = "com.project.springConfig")
 public class webConfig implements WebMvcConfigurer {
 
-    @Autowired
-    LocaleChangeInterceptor localeChangeInterceptor;
+    @Bean
+    public LocaleResolver localeResolver() {
+        SessionLocaleResolver slr = new SessionLocaleResolver();
+        slr.setDefaultLocale(Locale.US);
+        return slr;
+    }
+
+    @Bean
+    public LocaleChangeInterceptor localeChangeInterceptor() {
+        LocaleChangeInterceptor lci = new LocaleChangeInterceptor();
+        lci.setParamName("lang");
+        return lci;
+    }
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(localeChangeInterceptor);
+        registry.addInterceptor(localeChangeInterceptor());
     }
-
-
 }
