+++ src/main/resources/static/internationalization.js
@@ -1,0 +1,8 @@
+$(document).ready(function() {
+    $("#locales").change(function () {
+        var selectedOption = $('#locales').val();
+        if (selectedOption != ''){
+            window.location.replace('inter?lang=' + selectedOption);
+        }
+    });
+});
